Adds GNU License for public use
The GNU GPL is the most widely used free software license and has a
strong copyleft requirement. When distributing derived works, the
source code of the work must be made available under the same
license. There are multiple variants of the GNU GPL, each with
different requirements.

I allow the use of this software for commercial, modification,
distribution, and private use.