Adds use cases images for PHP and static files
I have been using this same image in multiple projects for more than
three years. I was tired of copying and pasting the same files all
over the place.

This image should give you an image that is ready for production and will
support your app during the bootstrap phase. There is enough buffer in
the configs to cover most of the PHP apps for enough time until you have
to concern yourself with performance and scalability.