Introduces FASTCGI_FILES_ROOT variable
The STATIC_FILES_ROOT variable is used to define the root dir for files
inside the nginx container.

In contrast, FASTCGI_FILES_ROOT, because the fcgi is another container,
needs to receive the information of where the files are hosted in the
fcgi container. The path gets appended with the script name from the
HTTP request and instructed to the fcgi server to execute.

Therefore, you can have different folder structure in each container.